Been attempting to ensure support for all varieties of table-layouts as well as reducing the time-complexity. This PR will solve #269 as well as open the door to supporting larger datasets and faster rendering.

Previously, when layout is being calculated, for each cell in the table it would traverse all cells that had come before it. It was not able to detect rowSpans left-wards beyond the immediate left. Not sure how to express this mathematically, but the time-complexity is/was exponential.

With this refactored table layout, the time-complexity is mostly reduced to match the complexity of the table. 

**The first commit includes failing tests.** One specifically for #269 and one generally that demonstrates an advanced implementation of rowSpan and colSpan.