@speedytwenty 2022-03-27
To briefly describe the refactored layout calculation mechanism (`layoutTable()`), it essentially allocates columns for rowSpans exceeding 1. I guess pseudo-code may best describe it concisely: 

```
function next(columnNumber) {
  if (columnNumber is allocated) return next(columnNumber + 1);
  return columnNumber;
}
table.forEach((row) => {
  row.forEach((cell) => {
    cell.x = next(currentColumn); // tracking of "currentColumn" not shown here
    if (cell.rowSpan > 1) allocate cell.rowSpan by cell.colSpan;
  });
  decrement all allocated columns by 1
});
```