@speedytwenty 2022-03-28
Here are some benchmarks that demonstrate a pretty substantial improvement in rendering performance:

## 5k cells

_1,000 rows X 5 columns_

before: 302.522ms
after: 169.297ms

## 50k cells

_10,000 rows X 5 columns_

before: 18.401s
after: 6.055s

## 100k cells

_10,000 rows X 10 columns_

before: 1:12.584 (m:ss.mmm)
after: 22.489s