@speedytwenty 2022-04-01
While there is still some refactoring left to do to enhance performance (#68, #285), this PR solves a bug as well as adds a test that demonstrates a really complex layout. To attempt to minimize confusion with continued refactoring and to get this fix out into the wild, I'm merging this now. Enjoy!